+++ lib/main.dart
@@ -146,6 +146,8 @@ class _RootAppWidgetState extends State<RootAppWidget>
         const Locale('hu', ''),
         const Locale('fr', ''),
         const Locale('es', ''),
+        const Locale('zh', ''),
+        const Locale('pt', ''), // currently BR?
       ],
       theme: Provider.of<ThemeProvider>(context).currentTheme(false),
       darkTheme: Provider.of<ThemeProvider>(context).currentTheme(true),