+++ lib/screens/nextcloud_auth_web_screen.dart
@@ -23,16 +23,18 @@ class _NextcloudAuthWebScreenState extends State<NextcloudAuthWebScreen> {
       ),
       body: Center(
         child: InAppWebView(
-          initialUrl: widget.authUrl,
-          initialHeaders: {
-            'OCS-APIRequest': 'true',
-          },
+          initialUrlRequest: URLRequest(
+            url: Uri.parse(widget.authUrl),
+            headers: {
+              'OCS-APIRequest': 'true',
+            },
+          ),
           initialOptions: InAppWebViewGroupOptions(
             crossPlatform: InAppWebViewOptions(
               useShouldOverrideUrlLoading: true,
               cacheEnabled: false,
               clearCache: true,
-              userAgent: 'Nextcloud Passwords App',
+              userAgent: 'Nextcloud Timemanager App',
               disableHorizontalScroll: true,
             ),
             android: AndroidInAppWebViewOptions(
@@ -43,11 +45,11 @@ class _NextcloudAuthWebScreenState extends State<NextcloudAuthWebScreen> {
           ),
           shouldOverrideUrlLoading:
               (controller, shouldOverrideUrlLoadingRequest) async {
-            final url = shouldOverrideUrlLoadingRequest.url;
             final uri1 = Uri.parse(widget.authUrl);
-            final uri2 = Uri.parse(url);
+            final uri2 = shouldOverrideUrlLoadingRequest.request.url;
+            final url = uri2.toString();
             if (!url.startsWith('nc://') && uri1.host != uri2.host) {
-              return ShouldOverrideUrlLoadingAction.CANCEL;
+              return NavigationActionPolicy.CANCEL;
             }
             if (url.startsWith('nc://login')) {
               final nap = Provider.of<NextcloudAuthProvider>(
@@ -56,7 +58,7 @@ class _NextcloudAuthWebScreenState extends State<NextcloudAuthWebScreen> {
               );
               await nap.setCredentials(url);
             }
-            return ShouldOverrideUrlLoadingAction.ALLOW;
+            return NavigationActionPolicy.ALLOW;
           },
           onLoadError: (controller, url, code, message) =>
               Navigator.of(context).pop(Provider.of<NextcloudAuthProvider>(
@@ -68,11 +70,9 @@ class _NextcloudAuthWebScreenState extends State<NextcloudAuthWebScreen> {
             context,
             listen: false,
           ).isAuthenticated),
-          onReceivedServerTrustAuthRequest: (InAppWebViewController controller,
-              ServerTrustChallenge challenge) async {
-            return ServerTrustAuthResponse(
-                action: ServerTrustAuthResponseAction.PROCEED);
-          },
+          onReceivedServerTrustAuthRequest: (controller, challenge) async =>
+              ServerTrustAuthResponse(
+                  action: ServerTrustAuthResponseAction.PROCEED),
         ),
       ),
     );