+++ plugins/autofill_service/lib/autofill_service.dart
@@ -13,7 +13,7 @@ enum AutofillServiceStatus {
 }
 
 class AutofillPreferences {
-  AutofillPreferences({this.enableDebug});
+  AutofillPreferences({this.enableDebug = false});
 
   factory AutofillPreferences.fromJson(Map<dynamic, dynamic> json) =>
       AutofillPreferences(enableDebug: json['enableDebug'] as bool);
@@ -39,11 +39,13 @@ class AutofillService {
     if (!Platform.isAndroid) {
       return false;
     }
-    return await _channel.invokeMethod('hasAutofillServicesSupport');
+    final r = await _channel.invokeMethod<bool>('hasAutofillServicesSupport');
+    return r != null && r;
   }
 
   Future<bool> get hasEnabledAutofillServices async {
-    return await hasAutofillServicesSupport && (await _channel.invokeMethod<bool>('hasEnabledAutofillServices'));
+    final r = await _channel.invokeMethod<bool>('hasAutofillServicesSupport');
+    return await hasAutofillServicesSupport && r != null && r;
   }
 
   Future<AutofillMetadata> getAutofillMetadata() async {
@@ -73,16 +75,19 @@ class AutofillService {
   }
 
   Future<bool> requestSetAutofillService() async {
-    return await _channel.invokeMethod('requestSetAutofillService');
+    final r = await _channel.invokeMethod<bool>('requestSetAutofillService');
+    return r != null && r;
   }
 
   Future<bool> resultWithDataset(
       {String label, String username, String password}) async {
-    return await _channel.invokeMethod('resultWithDataset', <String, dynamic>{
+    final r = await _channel.invokeMethod<bool>(
+        'resultWithDataset', <String, dynamic>{
       'label': label,
       'username': username,
       'password': password
     });
+    return r != null && r;
   }
 
   Future<void> disableAutofillServices() async {
@@ -105,26 +110,27 @@ class AutofillService {
 
 class AutofillMetadata {
   AutofillMetadata({this.packageNames, this.webDomains});
+
   factory AutofillMetadata.fromJson(Map<dynamic, dynamic> json) =>
       AutofillMetadata(
         packageNames: (json['packageNames'] as Iterable)
             .map((dynamic e) => e as String)
             .toSet(),
         webDomains: (json['webDomains'] as Iterable)
-            ?.map((dynamic e) =>
+            .map((dynamic e) =>
                 AutofillWebDomain.fromJson(e as Map<dynamic, dynamic>))
-            ?.toSet(),
+            .toSet(),
       );
 
   final Set<String> packageNames;
   final Set<AutofillWebDomain> webDomains;
 
   @override
-  String toString() => toJson()?.toString();
+  String toString() => toJson().toString();
 
   Map<String, Object> toJson() => {
         'packageNames': packageNames,
-        'webDomains': webDomains?.map((e) => e.toJson()),
+        'webDomains': webDomains.map((e) => e.toJson()),
       };
 }